+++ app/src/main/java/com/kabouzeid/gramophone/model/Album.java
@@ -21,7 +21,7 @@ public class Album implements Parcelable {
         this.songs = new ArrayList<>();
     }
 
-    public int getId() {
+    public long getId() {
         return safeGetFirstSong().albumId;
     }
 
@@ -29,7 +29,7 @@ public class Album implements Parcelable {
         return safeGetFirstSong().albumName;
     }
 
-    public int getArtistId() {
+    public long getArtistId() {
         return safeGetFirstSong().artistId;
     }
