+++ app/src/main/java/com/kabouzeid/gramophone/model/Song.java
@@ -9,19 +9,19 @@ import android.os.Parcelable;
 public class Song implements Parcelable {
     public static final Song EMPTY_SONG = new Song(-1, "", -1, -1, -1, "", -1, -1, "", -1, "");
 
-    public final int id;
+    public final long id;
     public final String title;
     public final int trackNumber;
     public final int year;
     public final long duration;
     public final String data;
     public final long dateModified;
-    public final int albumId;
+    public final long albumId;
     public final String albumName;
-    public final int artistId;
+    public final long artistId;
     public final String artistName;
 
-    public Song(int id, String title, int trackNumber, int year, long duration, String data, long dateModified, int albumId, String albumName, int artistId, String artistName) {
+    public Song(long id, String title, int trackNumber, int year, long duration, String data, long dateModified, long albumId, String albumName, long artistId, String artistName) {
         this.id = id;
         this.title = title;
         this.trackNumber = trackNumber;
@@ -59,16 +59,16 @@ public class Song implements Parcelable {
 
     @Override
     public int hashCode() {
-        int result = id;
+        int result = (int)id;
         result = 31 * result + (title != null ? title.hashCode() : 0);
         result = 31 * result + trackNumber;
         result = 31 * result + year;
         result = 31 * result + (int) (duration ^ (duration >>> 32));
         result = 31 * result + (data != null ? data.hashCode() : 0);
         result = 31 * result + (int) (dateModified ^ (dateModified >>> 32));
-        result = 31 * result + albumId;
+        result = 31 * result + (int)albumId;
         result = 31 * result + (albumName != null ? albumName.hashCode() : 0);
-        result = 31 * result + artistId;
+        result = 31 * result + (int)artistId;
         result = 31 * result + (artistName != null ? artistName.hashCode() : 0);
         return result;
     }
@@ -98,30 +98,30 @@ public class Song implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(this.id);
+        dest.writeLong(this.id);
         dest.writeString(this.title);
         dest.writeInt(this.trackNumber);
         dest.writeInt(this.year);
         dest.writeLong(this.duration);
         dest.writeString(this.data);
         dest.writeLong(this.dateModified);
-        dest.writeInt(this.albumId);
+        dest.writeLong(this.albumId);
         dest.writeString(this.albumName);
-        dest.writeInt(this.artistId);
+        dest.writeLong(this.artistId);
         dest.writeString(this.artistName);
     }
 
     protected Song(Parcel in) {
-        this.id = in.readInt();
+        this.id = in.readLong();
         this.title = in.readString();
         this.trackNumber = in.readInt();
         this.year = in.readInt();
         this.duration = in.readLong();
         this.data = in.readString();
         this.dateModified = in.readLong();
-        this.albumId = in.readInt();
+        this.albumId = in.readLong();
         this.albumName = in.readString();
-        this.artistId = in.readInt();
+        this.artistId = in.readLong();
         this.artistName = in.readString();
     }
