+++ app/src/main/java/com/kabouzeid/gramophone/ui/activities/ArtistDetailActivity.java
@@ -451,7 +451,7 @@ public class ArtistDetailActivity extends AbsSlidingMusicPanelActivity implement
 
     @Override
     public Loader<Artist> onCreateLoader(int id, Bundle args) {
-        return new AsyncArtistDataLoader(this, args.getInt(EXTRA_ARTIST_ID));
+        return new AsyncArtistDataLoader(this, args.getLong(EXTRA_ARTIST_ID));
     }
 
     @Override
@@ -467,9 +467,9 @@ public class ArtistDetailActivity extends AbsSlidingMusicPanelActivity implement
     }
 
     private static class AsyncArtistDataLoader extends WrappedAsyncTaskLoader<Artist> {
-        private final int artistId;
+        private final long artistId;
 
-        public AsyncArtistDataLoader(Context context, int artistId) {
+        public AsyncArtistDataLoader(Context context, long artistId) {
             super(context);
             this.artistId = artistId;
         }
